Move the Environment class to the utils module
This way also modules that do not depend on the ORT model have access to
basic Environment information, and Environment is now next to Os, which
is a sensible place.
