+++ reporter/src/main/kotlin/reporters/StaticHtmlReporter.kt
@@ -33,7 +33,6 @@ import kotlinx.html.*
 import kotlinx.html.dom.*
 
 import org.ossreviewtoolkit.downloader.VcsHost
-import org.ossreviewtoolkit.model.Environment
 import org.ossreviewtoolkit.model.Identifier
 import org.ossreviewtoolkit.model.Project
 import org.ossreviewtoolkit.model.Provenance
@@ -53,6 +52,7 @@ import org.ossreviewtoolkit.reporter.utils.ReportTableModel.ResolvableIssue
 import org.ossreviewtoolkit.reporter.utils.ReportTableModelMapper
 import org.ossreviewtoolkit.reporter.utils.SCOPE_EXCLUDE_LIST_COMPARATOR
 import org.ossreviewtoolkit.reporter.utils.containsUnresolved
+import org.ossreviewtoolkit.utils.Environment
 import org.ossreviewtoolkit.utils.ORT_FULL_NAME
 import org.ossreviewtoolkit.utils.isValidUri
 import org.ossreviewtoolkit.utils.normalizeLineBreaks
