+++ reporter/src/main/kotlin/utils/SpdxDocumentModelMapper.kt
@@ -22,7 +22,6 @@ package org.ossreviewtoolkit.reporter.utils
 import java.time.Instant
 import java.util.UUID
 
-import org.ossreviewtoolkit.model.Environment
 import org.ossreviewtoolkit.model.Identifier
 import org.ossreviewtoolkit.model.OrtResult
 import org.ossreviewtoolkit.model.ScanResult
@@ -41,6 +40,7 @@ import org.ossreviewtoolkit.spdx.model.SpdxExtractedLicenseInfo
 import org.ossreviewtoolkit.spdx.model.SpdxPackage
 import org.ossreviewtoolkit.spdx.model.SpdxPackageVerificationCode
 import org.ossreviewtoolkit.spdx.model.SpdxRelationship
+import org.ossreviewtoolkit.utils.Environment
 import org.ossreviewtoolkit.utils.ORT_FULL_NAME
 import org.ossreviewtoolkit.utils.ProcessedDeclaredLicense
 import org.ossreviewtoolkit.utils.stripCredentialsFromUrl
